Generalized app list builder.
