+++ apigee_edge.libraries.yml
@@ -34,13 +34,13 @@ apigee_edge.components:
     theme:
       css/apigee_edge.components.css: {}
 
-apigee_edge.listing:
+apigee_edge.app_listing:
   version: 1.0
   css:
     theme:
-      css/apigee_edge.listing.css: {}
+      css/apigee_edge.app_listing.css: {}
   js:
-    js/apigee_edge.listing.js: {}
+    js/apigee_edge.app_listing.js: {}
 
 apigee_edge.status_property:
   version: 1.0
